requirements.txt
streamlit
langchain-groq
langchain-core
PyPDF2
python-docx